Backend/iot-backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
requests==2.32.3
python-dotenv==1.0.1
